fix(analytics): documentation et correctif version standalone
