--- src/analytics/standalone/config.js
@@ -1,3 +1,0 @@
-import config from '../../core/config';
-config.namespace = 'dsfra';
-export default config;
