+++ src/analytics/standalone/example/index.ejs
@@ -5,7 +5,6 @@
     <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
     <meta name="format-detection" content="telephone=no,date=no,address=no,email=no,url=no">
     <title>Analytics standalone</title>
-    <link href="../component/connect/connect.standalone.css" rel="stylesheet" />
     <script>
       window['<%= namespace %>'] = {
         verbose: true,
