+++ src/core/script/api/options/modes.js
@@ -1,0 +1,9 @@
+export const Modes = {
+  AUTO: 'auto',
+  MANUAL: 'manual',
+  RUNTIME: 'runtime',
+  LOADED: 'loaded',
+  VUE: 'vue',
+  ANGULAR: 'angular',
+  REACT: 'react'
+};
